requirements.txt
# Core ML libraries
transformers>=4.40.0
torch>=2.0.0
datasets>=2.16.0
accelerate>=0.27.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
huggingface-hub>=0.20.0

# Evaluation and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Optional: for RLHF training
peft>=0.8.0  # Parameter Efficient Fine-Tuning
trl>=0.8.0   # Transformer Reinforcement Learning

